chore(tests): move metrics envtest to test/envtest
